feat: RAITA-34 Handle S3 duplicates.
Changed the field for the query from file_name to key, as it is surely unique.
Added another check for existingDoc to be sure that we are actually handling the
queried document, as opensearch query can return other relevant results as hits.
Added a mention to error message, that if it stems from -no index-, the index will
be created.