feat: Raita 34 handle s3 duplicates (#109)
* feat: RAITA-34 Upsert functionality for opensearch indexing, with temporary logging

* feat: RAITA-34 correct property for condition

* feat: RAITA-34 Handle S3 duplicates.

Removed console.logs, fixed the updateDoc body

* feat: RAITA-34 Handle S3 duplicates.

Changed the field for the query from file_name to key, as it is surely unique.
Added another check for existingDoc to be sure that we are actually handling the
queried document, as opensearch query can return other relevant results as hits.
Added a mention to error message, that if it stems from -no index-, the index will
be created.

* feat: RAITA-34 Handle S3 duplicates.

Updated the query to match the keyword of file_name and key fields.
Refactored most of the functionality from searchExisting to upsertDocument
searchExisting now only does the search and returns the result what ever it
will be.

* feat: RAITA-34 Handle S3 Duplicates

Refactored the upsertDocument to return the promises to saveFileMetaData as <void>
because at the moment we don't use the returning values in anything else, but
error handling. file.fileBody was stringified in two places, so refactored that
also.